log, image kept when email failed

+++ server/api/actionform.php
@@ -8,8 +8,13 @@ use PHPMailer\PHPMailer\Exception;
 require '../vendor/phpmailer/phpmailer/src/Exception.php';
 require '../vendor/phpmailer/phpmailer/src/PHPMailer.php';
 require '../vendor/phpmailer/phpmailer/src/SMTP.php';
+use \DateTime;
 
 $full_debug = false;
+$action_saved = false;
+$action_mailed = false;
+
+
 
 if ($full_debug) {
     ini_set('display_errors', 1);
@@ -24,7 +29,7 @@ require  __DIR__ . '/actionsave.php';
 
 $kopernik_config = parse_ini_file('kopernik.ini');
 
-function prepare_new_name($old_name){
+function prepare_new_name($old_name) {
 
     $fileNameCmps = explode(".", $old_name);
     $fileExtension = strtolower(end($fileNameCmps));
@@ -36,17 +41,42 @@ function prepare_new_name($old_name){
     return '';
 }
 
-$uploaded_image = $_FILES['image']['tmp_name'];
+var_dump($_FILES);
+
+$uploaded_image_source = $_FILES['image']['tmp_name'];
 $uploaded_image_name_old = basename($_FILES["image"]["name"]);
 $uploaded_image_name_new = $_POST['scoutunit'] . '_' . prepare_new_name($uploaded_image_name_old);
 $uploaded_image_size = $_FILES['image']['size'];
+$uploaded_image = "../evidence" . $uploaded_image_name_new;
+if(is_uploaded_file($uploaded_image_source)){
+    move_uploaded_file($uploaded_image_source,  $uploaded_image);
+}
 // extra checks https://www.php.net/manual/en/features.file-upload.php
 
+function oneliner($text){
+    return str_replace(PHP_EOL, "; ",  $text);
+}
+
+function log_base($text) {
+    $timestamp = new DateTime();
+    file_put_contents('../logs/log_'.date("Y-m-d").'.txt', $timestamp->format(DateTime::ATOM) ."\t" . $text . PHP_EOL, FILE_APPEND | LOCK_EX);
+}
+
+function log_info($text){
+    log_base( "INFO  " . oneliner($text) . PHP_EOL );
+}
+
+function log_error($error){
+    $error_start = "ERROR "; 
+    $errlines = $error_start . str_replace(PHP_EOL , PHP_EOL . $error_start, $error);
+    log_base($errlines);
+}
 
 function cleanup(){
     global $uploaded_image;
+    global $action_mailed;
 
-    if (file_exists($uploaded_image)) {
+    if (file_exists($uploaded_image) && $action_mailed) {
         unlink($uploaded_image);
      }
 }
@@ -56,6 +86,7 @@ function redirect_to_index()
     cleanup();
     header('Location: ../index.html');
 }
+
 function redirect_to_result()
 {   
     cleanup();
@@ -67,6 +98,8 @@ function clean_field($input)
     $sanitized = trim( htmlentities( $input, ENT_QUOTES, 'UTF-8') );
     return (str_contains($sanitized,';') ? '' : $sanitized);
 }
+
+
 
 
 // $inputs = [
@@ -126,33 +159,22 @@ echo "image new name=". $uploaded_image_name_new . PHP_EOL;
 
 
 
-if($uploaded_image_size==0 || $uploaded_image_size>8000000 || !$uploaded_image_name_new ||  !is_uploaded_file($uploaded_image)) 
+if($uploaded_image_size==0 || $uploaded_image_size>8000000 || !$uploaded_image_name_new ||  !file_exists($uploaded_image)) 
 {
     array_push($errors, 'image');   
 }
 
 if($errors)
 {
+    log_error("VALIDATION " . $errors);
     redirect_to_result();
     return;
 }
 
-redirect_to_result();
+if($full_debug){
 
-$mysqli = kopernik_mysqli($kopernik_config, $full_debug);
-
-$action_saved = false;
-if($mysqli->connect_errno == 0)
-{
-    $err = save_action($mysqli, $data);
-
-    if (!$err) {
-        $action_saved = true;
-    }
 } else {
-    // if (mysqli_connect_errno()) {
-    //     throw new RuntimeException('mysqli connection error: ' . mysqli_connect_error());
-    // }
+    redirect_to_result();
 }
 
 
@@ -167,12 +189,31 @@ foreach($data as $key=>$value)
     $mail_body .= $key . '=' . $value . PHP_EOL;
 }
 $mail_body .= 'image_old' . '=' . $uploaded_image_name_old . PHP_EOL;
+
+
+$mysqli = kopernik_mysqli($kopernik_config, $full_debug);
+
+if($mysqli->connect_errno == 0)
+{
+    $err = save_action($mysqli, $data);
+
+    if (!$err) {
+        $action_saved = true;
+    } else {
+        log_error($err . PHP_EOL . oneliner($mail_body));
+    }
+} else {
+    log_error(mysqli_connect_error() . PHP_EOL . oneliner($mail_body));
+    // if (mysqli_connect_errno()) {
+    //     throw new RuntimeException('mysqli connection error: ' . mysqli_connect_error());
+    // }
+}
+
 if($action_saved)
 {
     $mail_body .= 'action_saved' . '=' . 'true' . PHP_EOL;
 } else {
     $mail_body .= 'action_saved' . '=' . 'false' . PHP_EOL;
-
 }
 
 //var_dump($mail_body);
@@ -218,8 +259,10 @@ try {
 
     $mail->send();
     echo 'Message has been sent';
+    log_info("EMAIL=SENT". PHP_EOL . $mail_body);
+
 } catch (Exception $e) {
-    echo "Message could not be sent. Mailer Error: {$mail->ErrorInfo}";
+    log_error("EMAIL=FAIL; {$mail->ErrorInfo}");
 }
